+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 - PR #4849 Update Java bindings to use new NVTX API
 - PR #4845 Add CUDF_FUNC_RANGE to top-level cuIO function APIs
 - PR #4848 Side step `unique_count` calculation in `scatter_by_map`
+- PR #4870 Add chunked parquet file writing from python
 
 ## Bug Fixes